Fixes ray casting, adds simple turnable camera
Fixes a simple degrees vs radians bug in the ray casting implementation.
Also fixes the structure of the box to be a box (one of the edges was
going through the center instead of along the side)

The camera can be turned by pressing a and d (or left arrow and right
arrow).